feat: [dijkstra] - rename module
Changes
---

최단경로 내 3가지 알고리즘 중 다익스트라
