+++ ShortestPath/README.md
@@ -1,18 +1,18 @@
-# Dijkstra Algorithm
+# 최단경로
 
-### 출발점에서 목표점까지의 ***최단 경로***를 구하는 알고리즘
+## 1. Dijkstra Algorithm
+#### :출발점에서 목표점까지의 ***최단 경로***를 구하는 알고리즘
 
-## 특징
+### 특징
 - 한 노드에서 다른 모든 노드로의 최단 경로를 구할 수 있음
 - 단, edge에 0보다 작은 weight가 없어야 함
 - **_Greedy + DP_**
 - 복잡도: O(ElogV)
 
-##
-### main
+####
+### 구현
 - [x] 다익스트라 기본 구현(한 노드에서 다른 모든 노드까지의 각 최단 경로)
 - [x] 우선순위큐로 구현
-### practice
